Move ChannelList class to separate file
